+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New process to add fulfill urls to Limited Access manifests and update fulfill_limited_access flags to True
 - Updated README and added more information to installation steps
 - Deprecated datetime.utcnow() method
+- Added new column (contract_source) to the records and item tables
 ## Fixed
 - Resolved the format of fulfill endpoints in UofM manifests